+++ changelog.md
@@ -5,7 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added 
+- The ability to upload historical data which is now displayed alongside the projected data in the charts.
+- A warning that appears if the manually-entered census value from the sidebar does not match the census value provided in the actuals.
+- A section at the bottom of the page with release notes and changes by version.
+### Changed
+- The "Hospital Capacity" section header to "COVID-19 Hospital Capacity".
+### Fixed
+- Hopefully fixed the tooltip bug where the tooltip was showing the date value as one day behind the actual uploaded/calculated date values.
 
 ## [1.3.3] - 2020-04-03
 ### Fixed